# [clang] Optimize EmbedExpr child iterator callback